Move examples to their own lib and import them to the viewer bin
