+++ crates/viewer/Cargo.toml
@@ -5,8 +5,9 @@ edition = "2021"
 
 [dependencies]
 bytemuck = { version = "1", features = ["derive"] }
-opencascade = { version = "0.1", path = "../opencascade" }
+examples = { path = "../../examples" }
 glam = { version = "0.23", features = ["bytemuck"] }
+opencascade = { version = "0.1", path = "../opencascade" }
 simple-game = { git = "https://github.com/bschwind/simple-game.git" }
 smaa = "0.10"
 wgpu = "0.16"
