+++ examples/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "examples"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+opencascade = { version = "0.1", path = "../crates/opencascade" }
+glam = { version = "0.23", features = ["bytemuck"] }
