+++ examples/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod box_shape;
+pub mod chamfer;
+pub mod gizmo;
+pub mod high_level_bottle;
+pub mod keyboard_case;
+pub mod keycap;
+pub mod rounded_chamfer;
