+++ examples/src/rounded_chamfer.rs
@@ -1,10 +1,13 @@
 use glam::dvec3;
-use opencascade::{primitives::Direction, workplane::Workplane};
+use opencascade::{
+    primitives::{Direction, Shape},
+    workplane::Workplane,
+};
 
 // Demonstrates filleting a 2D profile, extruding it, then chamfering
 // the top edges, resulting in a nice, rounded chamfer.
 
-pub fn main() {
+pub fn shape() -> Shape {
     let mut base = Workplane::xy().rect(16.0, 10.0);
     base.fillet(1.0);
     let shape = base.to_face().extrude(dvec3(0.0, 0.0, 3.0));
@@ -13,5 +16,5 @@ pub fn main() {
     let top_edges = shape.faces().farthest(Direction::PosZ).edges();
     shape.chamfer_edges(0.7, top_edges);
 
-    shape.write_stl("rounded_chamfer.stl").unwrap();
+    shape
 }
